requirements.txt
pydantic-ai-slim[openai,logfire]==0.0.28
openai
black